v1/backend/requirements.txt
fastapi
uvicorn
pytesseract
pillow
httpx
python-dotenv
python-multipart